TopazUtils追加
