+++ Packages/com.nomlas.topazchat/Runtime/Scripts/TopazUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90f0a5b477a760b4982d3fe88d1ddc2e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
